+++ src/style/mod.rs
@@ -16,17 +16,19 @@
 use embedded_graphics::{prelude::*, style::TextStyle};
 
 pub mod builder;
+pub mod height_mode;
 
 use crate::{
     alignment::{HorizontalTextAlignment, VerticalTextAlignment},
     parser::{Parser, Token},
+    style::height_mode::HeightMode,
     utils::font_ext::FontExt,
 };
 pub use builder::TextBoxStyleBuilder;
 
 /// Styling options of a [`TextBox`].
 ///
-/// `TextBoxStyle` contains the `Font`, foreground and background `PixelColor`,
+/// `TextBoxStyle` contains the `Font`, foreground and background `PixelColor`, [`HeightMode`],
 /// [`HorizontalTextAlignment`] and [`VerticalTextAlignment`] information necessary to draw a
 /// [`TextBox`].
 ///
@@ -40,12 +42,13 @@ pub use builder::TextBoxStyleBuilder;
 /// [`new`]: #method.new
 /// [`from_text_style`]: #method.from_text_style
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Default)]
-pub struct TextBoxStyle<C, F, A, V>
+pub struct TextBoxStyle<C, F, A, V, H>
 where
     C: PixelColor,
     F: Font + Copy,
     A: HorizontalTextAlignment,
     V: VerticalTextAlignment,
+    H: HeightMode,
 {
     /// Style properties for text.
     pub text_style: TextStyle<C, F>,
@@ -55,22 +58,33 @@ where
 
     /// Vertical text alignment.
     pub vertical_alignment: V,
+
+    /// The height behaviour
+    pub height_mode: H,
 }
 
-impl<C, F, A, V> TextBoxStyle<C, F, A, V>
+impl<C, F, A, V, H> TextBoxStyle<C, F, A, V, H>
 where
     C: PixelColor,
     F: Font + Copy,
     A: HorizontalTextAlignment,
     V: VerticalTextAlignment,
+    H: HeightMode,
 {
     /// Creates a `TextBoxStyle` object with transparent background.
     #[inline]
-    pub fn new(font: F, text_color: C, alignment: A, vertical_alignment: V) -> Self {
+    pub fn new(
+        font: F,
+        text_color: C,
+        alignment: A,
+        vertical_alignment: V,
+        height_mode: H,
+    ) -> Self {
         Self {
             text_style: TextStyle::new(font, text_color),
             alignment,
             vertical_alignment,
+            height_mode,
         }
     }
 
@@ -80,11 +94,13 @@ where
         text_style: TextStyle<C, F>,
         alignment: A,
         vertical_alignment: V,
+        height_mode: H,
     ) -> Self {
         Self {
             text_style,
             alignment,
             vertical_alignment,
+            height_mode,
         }
     }